+++ earthdaily/earthdatastore/cube_utils/__init__.py
@@ -3,7 +3,10 @@ import numpy as np
 import pandas as pd
 import geopandas as gpd
 from shapely.geometry import box
-from earthdaily.earthdatastore.cube_utils._zonal import zonal_stats, zonal_stats_numpy
+from earthdaily.earthdatastore.cube_utils._zonal import (
+    zonal_stats,
+    zonal_stats_numpy,
+)
 from rasterio.enums import Resampling
 from rasterio.mask import geometry_mask
 import rioxarray as rxr
@@ -24,10 +27,14 @@ def _apply_nodata(ds, nodata_assets: dict):
 
 
 def _autofix_unfrozen_coords_dtype(ds):
-    attrs = {c: ds.coords[c].data.tolist() for c in ds.coords if c not in ds.dims}
+    attrs = {
+        c: ds.coords[c].data.tolist() for c in ds.coords if c not in ds.dims
+    }
     # force str
     for attr in attrs:
-        if not isinstance(attrs[attr], (str, int, float, np.ndarray, list, tuple)):
+        if not isinstance(
+            attrs[attr], (str, int, float, np.ndarray, list, tuple)
+        ):
             ds.coords[attr] = str(attrs[attr])
             ds.coords[attr] = ds.coords[attr].astype(str)
     return ds
@@ -129,7 +136,9 @@ def datacube(
                 continue
             nodata = (
                 item.assets[asset]
-                .extra_fields.get("raster:bands", empty_dict_list)[band_idx - 1]
+                .extra_fields.get("raster:bands", empty_dict_list)[
+                    band_idx - 1
+                ]
                 .get("nodata")
             )
             if nodata == 0 or nodata:
@@ -162,9 +171,9 @@ def datacube(
             invert=True,
         )
 
-        mask_ = xr.DataArray(data=clip_mask_arr, coords=dict(y=ds.y, x=ds.x)).chunk(
-            chunks=dict(x=ds.chunks["x"][0], y=ds.chunks["y"][0])
-        )
+        mask_ = xr.DataArray(
+            data=clip_mask_arr, coords=dict(y=ds.y, x=ds.x)
+        ).chunk(chunks=dict(x=ds.chunks["x"][0], y=ds.chunks["y"][0]))
 
         ds = ds.where(mask_)
         del mask_, clip_mask_arr
@@ -197,7 +206,9 @@ def rescale_assets_with_items(
             if item.datetime in unique_dt.keys():
                 for asset in item.assets.keys():
                     if asset not in unique_dt[item.datetime].assets.keys():
-                        unique_dt[item.datetime].assets[asset] = item.assets[asset]
+                        unique_dt[item.datetime].assets[asset] = item.assets[
+                            asset
+                        ]
             else:
                 unique_dt[item.datetime] = item
             # items_collection_unique_dt.append(item)
@@ -209,9 +220,9 @@ def rescale_assets_with_items(
 
     for idx, time in enumerate(ds.time.values):
         current_item = items_collection[idx]
-        if pd.Timestamp(time).strftime("%Y%m%d") != current_item.datetime.strftime(
+        if pd.Timestamp(time).strftime(
             "%Y%m%d"
-        ):
+        ) != current_item.datetime.strftime("%Y%m%d"):
             raise ValueError(
                 "Mismatch between items and datacube, cannot scale data. Please set rescale to False."
             )
@@ -285,7 +296,10 @@ def rescale_assets_with_items(
                     times = list(set(scales[asset][scale][offset]))
                     if len(times) != len(scales[asset][scale][offset]):
                         for time in times:
-                            d = ds[[asset]].loc[dict(time=time)] * scale + offset
+                            d = (
+                                ds[[asset]].loc[dict(time=time)] * scale
+                                + offset
+                            )
                             ds_scaled[asset].append(d)
                     else:
                         d = ds[[asset]].loc[dict(time=times)] * scale + offset
@@ -324,7 +338,9 @@ def _drop_unfrozen_coords(ds):
 
 
 def _common_data_vars(*cubes):
-    data_vars = list(set([k for cube in cubes for k in list(cube.data_vars.keys())]))
+    data_vars = list(
+        set([k for cube in cubes for k in list(cube.data_vars.keys())])
+    )
     return data_vars
 
 
@@ -361,6 +377,8 @@ def metacube(*cubes, concat_dim="time", by="time.date", how="mean"):
                     cubes[idx][data_var] == np.nan, other=np.nan
                 )
 
-    cube = xr.concat([_drop_unfrozen_coords(cube) for cube in cubes], dim=concat_dim)
+    cube = xr.concat(
+        [_drop_unfrozen_coords(cube) for cube in cubes], dim=concat_dim
+    )
     cube = _groupby(cube, by=by, how=how)
     return _propagade_rio(cubes[0], cube)