+++ examples/field_evolution.py
@@ -27,30 +27,27 @@ pivot = gpd.read_file("pivot.geojson")
 eds = earthdatastore.Auth()
 
 ##############################################################################
-# Search for collection items in august 2022
-#
+# Search for collection items in august 2022 (1st to 9th)
+# where at least 50% of the field is clear according to the native cloudmask.
+
 pivot_cube = eds.datacube(
     "sentinel-2-l2a",
     intersects=pivot,
-    datetime=["2022-08-01", "2022-08-08"],
+    datetime=["2022-08-01", "2022-08-09"],
     assets=["red", "green", "blue"],
     mask_with="native",  # same as scl
-    mask_statistics=True,
+    # mask_statistics=True, # as you ask `clear_cover`it will force computing mask_statistics
+    clear_cover=50,
 )
 
-pivot_cube.clear_percent_scl.plot.scatter(x="time")
+pivot_cube.clear_percent.plot.scatter(x="time")
 
-#####################################################################da#########
+##############################################################################
 # Plots cube with SCL with at least 50% of clear data
 # ----------------------------------------------------
-cube_majority_clear = pivot_cube.sel(
-    time=pivot_cube.time[pivot_cube.clear_percent_scl > 50]
-)
 
-cube_majority_clear = cube_majority_clear.load()
-
-cube_majority_clear.to_array(dim="band").plot.imshow(
-    vmin=0, vmax=0.33, col="time", col_wrap=3
+pivot_cube.to_array(dim="band").plot.imshow(
+    vmin=0, vmax=0.4, col="time", col_wrap=3
 )
 
 plt.title("Clear cover percent with SCL")
@@ -58,12 +55,12 @@ plt.title("Pivot evolution with SCL masks")
 plt.show()
 
 
-#####################################################################da#########
+##############################################################################
 # Compute zonal stats for the pivot
 # ----------------------------------------------------
 
 zonal_stats = earthdatastore.cube_utils.zonal_stats(
-    cube_majority_clear, pivot, operations=["mean", "max", "min"]
+    pivot_cube, pivot, operations=["mean", "max", "min"]
 )
 zonal_stats = zonal_stats.load()